refactor generate-svgs script as function to import in plugin
